final cmmit day 10

+++ day 10 function.py
@@ -57,14 +57,14 @@ operation= {
 }
 
 def cal():
-    num1=int(input("first number: "))
+    num1=float(input("first number: "))
     for symbol in operation:
       print(symbol)
     should_con=True
     while should_con:
      operation_symbol=input("Pick the symbol: ")
      cal_function=operation[operation_symbol]
-     num2=int(input("secound number: "))
+     num2=float(input("secound number: "))
      answer=cal_function(num1,num2)
 
      print(f"{num1} {operation_symbol} {num2} ={answer}")
